✨ feat: 통계용 함수 추가 통계테이블 엔티티 추가

+++ backend/focus/src/main/java/com/bb/focus/db/entity/statistics/MajorPerApplicant.java
@@ -34,6 +34,6 @@ public class MajorPerApplicant {
 
   @NotNull
   @ColumnDefault("0")
-  private Long applicantCount;
+  private int applicantCount;
 
 }
